+++ CHANGELOG.md
@@ -22,6 +22,9 @@
 
 #### Enhancements
 
+* Adds `empty_string` opt-in rule.
+[Davide Sibilio](https://github.com/idevid)
+
 * Adds `discouraged_optional_boolean` opt-in rule to discourage
   the use of optional booleans.  
   [Ornithologist Coder](https://github.com/ornithocoder)